requirements.txt
torch>=1.7
numpy
matplotlib
scikit-learn
nibabel
Pillow
opencv-python